+++ helpscout/__init__.py
@@ -1,4 +1,4 @@
 from helpscout.client import HelpScout  # noqa
 
 
-__version__ = '0.2.3'
+__version__ = '1.0.0'